+++ processor/src/test/resources/expected/com/example/arez/React4j_ComponentWithCallback.java
@@ -38,11 +38,6 @@ abstract class React4j_ComponentWithCallback extends ComponentWithCallback {
     return componentConstructor;
   }
 
-  @Override
-  protected boolean shouldComponentUpdate(@Nullable final JsPropertyMap<Object> nextProps) {
-    return false;
-  }
-
   @Nonnull
   private Callback.Procedure create_handleFoo() {
     final Callback.Procedure handler = () -> this.handleFoo();