+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 * **\[arez\]** Replace `reportPropsChanged(...)` with `shouldComponentUpdate(...)` on the `ReactArezComponent`
   class so that subclasses can control whether a prop change should update component rather than just reporting
   a prop change as an arez change.
+* **\[processor\]** Avoid generating an empty `shouldComponentUpdate(...)` if there are no props to check and
+  instead add a concrete implementation on the `ReactArezComponent` class. This reduces the code size of compiled
+  output.
 
 ### [v0.88](https://github.com/react4j/react4j/tree/v0.88) (2018-07-26)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.87...v0.88)